Finalize Rails5 release

+++ app/controllers/org_admin/question_options_controller.rb
@@ -8,7 +8,7 @@ module OrgAdmin
 
     after_action :verify_authorized
 
-
+    # rubocop:disable Metrics/AbcSize
     def destroy
       question_option = QuestionOption.find(params[:id])
       option_id_to_remove = question_option.id.to_s
@@ -20,15 +20,13 @@ module OrgAdmin
         if question_option.destroy!
           # need to remove any conditions which refer to this question option
           question.conditions.each do |cond|
-            if cond.option_list.include?(option_id_to_remove)
-              cond.destroy
-            end
+            cond.destroy if cond.option_list.include?(option_id_to_remove)
           end
           flash[:notice] = success_message(question_option, _("deleted"))
         else
           flash[:alert] = flash[:alert] = failure_message(question_option, _("delete"))
         end
-      rescue StandardError => e
+      rescue StandardError
         flash[:alert] = _("Unable to create a new version of this template.")
       end
       redirect_to edit_org_admin_template_phase_path(
@@ -37,8 +35,8 @@ module OrgAdmin
         section: section.id
       )
     end
+    # rubocop:enable Metrics/AbcSize
 
   end
-
 
 end
